+++ src/Base/AbstractBase.php
@@ -1,7 +1,11 @@
 <?php
+
+declare(strict_types=1);
+
 namespace JiNexus\Route\Base;
 
-use JiNexus\Route\Exception;
+use JiNexus\Route\RouteException;
+use ReflectionException;
 
 /**
  * Class AbstractBase
@@ -12,17 +16,17 @@ abstract class AbstractBase implements BaseInterface
     /**
      * Base setters and getters
      *
-     * @param $name
+     * @param $property
      * @param array $arguments
      * @return $this|mixed
-     * @throws Exception
+     * @throws RouteException | ReflectionException
      */
-    public function __call($name, array $arguments)
+    public function __call($property, array $arguments)
     {
-        $action = substr($name, 0, 3);
+        $action = substr($property, 0, 3);
 
         if ( $action == 'get' || $action == 'set' ) {
-            $property = lcfirst(substr($name, 3));
+            $property = lcfirst(substr($property, 3));
 
             if ( property_exists($this, $property) ) {
                 $reflection = new \ReflectionObject($this);
@@ -39,6 +43,6 @@ abstract class AbstractBase implements BaseInterface
             }
         }
 
-        throw new Exception('Not implemented: ' . get_called_class() . '::' . $name);
+        throw new RouteException('Not implemented: ' . get_called_class() . '::' . $property);
     }
 }