+++ src/Redirect/Factory/RedirectFactory.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace JiNexus\Route\Redirect\Factory;
 
 use JiNexus\Route\Factory\AbstractFactory;
@@ -13,7 +16,7 @@ class RedirectFactory extends AbstractFactory
     /**
      * @return Redirect
      */
-    public static function build()
+    public static function build(): Redirect
     {
         return new Redirect();
     }