+++ src/RouteException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JiNexus\Route;
+
+/**
+ * Exception class
+ */
+class RouteException extends \Exception
+{ }